Update README.md
# Next.js Project Setup and Basics

Welcome to the **alx-project-0x02-setup** repository!

## Introduction

This repository serves as a foundational setup for a Next.js project. It is part of the ALX Software Engineering curriculum and is designed to introduce learners to the basic setup and configuration required to get started with Next.js.

The main goals of this project include:

- Understanding the folder structure of a Next.js app
- Setting up a working development environment
- Creating pages and navigating between them
- Exploring basic features like routing, components, and static file serving

This setup will provide a solid base for building more advanced web applications using Next.js.